Resolving few more comments
